requirements.txt
streamlit
python-dotenv
langdetect
Pillow
google-ai-generativelanguage
crewai
faster-whisper
whisper
soundfile
pydub
